requirements.txt
playwright>=1.41.0
beautifulsoup4>=4.12.0
openai>=1.10.0
pydantic>=2.6.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-google-genai>=0.0.5
pandas>=2.2.0
openpyxl>=3.1.0
